Shard the Compute test to fix the new slowest path
With publiclibrary.js already split, functions.js's "Compute" test
(110 operators x 3375 argument combinations each = 371k evaluations)
became the dominant bottleneck at ~8 minutes while every other shard
finished in 1-2 minutes. Split it into compute-1..4.js (~27-28
operators each, sharing the loop logic via compute-util.js) and folded
the quick "Dynamic expressions" test into compute-4.js instead of
giving it its own job. Updated both workflow matrices accordingly.
