+++ tests/testcafe/compute-3.js
@@ -1,0 +1,7 @@
+import { compute_ops } from '../../client/js/compute.js';
+import { setupTestEnvironment } from './test-util.js';
+import { computeTest } from './compute-util.js';
+
+setupTestEnvironment();
+
+computeTest(compute_ops.slice(56, 83), 'ops 57-83');
